Refactor: extract auth endpoints to accountController
